+++ plugin_ide.core/src/com/googlecode/goclipse/core/operations/GoBuildManager.java
@@ -13,13 +13,13 @@ package com.googlecode.goclipse.core.operations;
 import static melnorme.lang.ide.core.utils.ResourceUtils.loc;
 
 import java.nio.file.Path;
-import java.text.MessageFormat;
 import java.util.Collection;
 
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
 
+import com.googlecode.goclipse.core.GoCoreMessages;
 import com.googlecode.goclipse.core.GoEnvironmentPrefs;
 import com.googlecode.goclipse.core.GoProjectEnvironment;
 import com.googlecode.goclipse.tooling.GoBuildOutputProcessor;
@@ -32,23 +32,46 @@ import melnorme.lang.ide.core.operations.OperationInfo;
 import melnorme.lang.ide.core.operations.ToolMarkersUtil;
 import melnorme.lang.ide.core.operations.build.BuildManager;
 import melnorme.lang.ide.core.operations.build.BuildTarget;
+import melnorme.lang.ide.core.operations.build.BuildTarget.BuildType;
 import melnorme.lang.ide.core.operations.build.CommonBuildTargetOperation;
 import melnorme.lang.ide.core.project_model.AbstractBundleInfo;
-import melnorme.lang.ide.core.project_model.AbstractBundleInfo.BuildConfiguration;
 import melnorme.lang.ide.core.project_model.LangBundleModel;
 import melnorme.lang.ide.core.project_model.ProjectBuildInfo;
 import melnorme.lang.ide.core.utils.ResourceUtils;
 import melnorme.lang.tooling.data.StatusLevel;
 import melnorme.utilbox.collections.ArrayList2;
+import melnorme.utilbox.collections.Indexable;
 import melnorme.utilbox.concurrency.OperationCancellation;
 import melnorme.utilbox.core.CommonException;
 import melnorme.utilbox.misc.Location;
+import melnorme.utilbox.misc.MiscUtil;
 import melnorme.utilbox.process.ExternalProcessHelper.ExternalProcessResult;
 
 public class GoBuildManager extends BuildManager {
 	
 	public GoBuildManager(LangBundleModel<? extends AbstractBundleInfo> bundleModel) {
 		super(bundleModel);
+	}
+	
+	public static final String BUILD_TYPE_Build = "build";
+	public static final String BUILD_TYPE_BuildTests = "build-tests";
+	public static final String BUILD_TYPE_RunTests = "[run-tests]";
+	
+	public static final Indexable<BuildType> BUILD_TYPES = createDefaultBuildTypes();
+	public static final Indexable<String> BUILD_TYPES_Names = 
+			BUILD_TYPES.map((buildType) -> buildType.getName());
+	
+	public static ArrayList2<BuildType> createDefaultBuildTypes() {
+		return ArrayList2.create(
+			new GoDefaultBuildType(),
+			new GoTestBuildType(),
+			new GoRunTestsBuildType()
+		);
+	}
+	
+	@Override
+	protected Indexable<BuildType> getBuildTypes_do() {
+		return BUILD_TYPES;
 	}
 	
 	@Override
@@ -58,49 +81,130 @@ public class GoBuildManager extends BuildManager {
 			return buildTarget;
 		}
 		
-		IProject project = projectBuildInfo.getProject();
-		GoEnvironment goEnv = GoProjectEnvironment.getGoEnvironment(project);
-		
-		GoPackageName goPackage = GoPackageName.createValid(targetName);
-		Location goPackageFile = goEnv.getBinFileForGoPackage(loc(project.getLocation()), goPackage);
-		
-		BuildConfiguration buildConfig = new BuildConfiguration(targetName, goPackageFile.toPath());
-		return createBuildTarget(targetName, buildConfig, false, "");
-	}
-	
-	@Override
-	public String getDefaultBuildOptions(String targetName) {
-		/* FIXME: todo according to targetName*/
-		return "install -v -gcflags \"-N -l\" ./...";
+		return createBuildTarget(targetName, false, null);
 	}
 	
 	@Override
 	public CommonBuildTargetOperation createBuildTargetSubOperation(OperationInfo parentOpInfo, IProject project,
 			Path buildToolPath, BuildTarget buildTarget, boolean fullBuild) {
-		return new GoBuildTargetOperation(parentOpInfo, project, buildToolPath, buildTarget, fullBuild);
+		return new GoBuildTargetOperation(this, parentOpInfo, project, buildToolPath, buildTarget, fullBuild);
 	}
 	
 	/* -----------------  ----------------- */
 	
-	protected class GoBuildTargetOperation extends CommonBuildTargetOperation {
+	public static abstract class AbstractGoBuildType extends BuildType {
 		
-		protected static final String ERROR_SrcRootContainsGoFiles = 
-				"The Go `src` directory at `{0}` contains .go files. " +
-				"This is not allowed, these files will be ignored.\n" + 
-				"Instead, all .go files should be in a subdirectory of `src`, " + 
-				"so that they will be part of a Go package. " + 
-				"This is so they can be built using the `./...` pattern, or imported by other Go files.";
+		public AbstractGoBuildType(String name) {
+			super(name);
+		}
+		
+		protected GoPackageName getValidGoPackageName(String goPackageString) throws CommonException {
+			return GoPackageName.createValid(goPackageString);
+		}
+		
+		@Override
+		public String getDefaultBuildOptions(BuildTarget buildTarget, IProject project) throws CommonException {
+			String goPackageSpec = getGoPackageSpec(project, buildTarget.getBuildConfiguration());
+			return getBuildCommand() + " -v -gcflags \"-N -l\" " + goPackageSpec;
+		}
+		
+		protected abstract String getBuildCommand();
+		
+		protected String getGoPackageSpec(IProject project, String goPackageSpec) throws CommonException {
+			
+			if(goPackageSpec == null || goPackageSpec.isEmpty()) {
+				GoEnvironment goEnv = GoProjectEnvironment.getGoEnvironment(project);
+				GoPackageName goPackage = goEnv.getGoPath().findGoPackageForLocation(loc(project.getLocation()));
+				
+				if(goPackage != null) {
+					goPackageSpec = goPackage.toString();
+				} else {
+					goPackageSpec = ".";
+				}
+				
+				goPackageSpec += "/...";
+			}
+			return goPackageSpec;
+		}
+		
+		@Override
+		public Path getArtifactPath(BuildTarget buildTarget, IProject project) throws CommonException {
+			Location binFolderLocation = GoProjectEnvironment.getBinFolderLocation(project);
+			
+			String binFilePath = getBinFilePath(getValidGoPackageName(buildTarget.getBuildConfiguration()));
+			return binFolderLocation.resolve(binFilePath + MiscUtil.getExecutableSuffix()).toPath();
+		}
+		
+		protected String getBinFilePath(GoPackageName goPackageName) throws CommonException {
+			return goPackageName.getLastSegment();
+		}
+		
+	}
+	
+	public static class GoDefaultBuildType extends AbstractGoBuildType {
+		
+		public GoDefaultBuildType() {
+			super(BUILD_TYPE_Build);
+		}
+		
+		@Override
+		protected String getBuildCommand() {
+			return "install";
+		}
+		
+	}
+	
+	public static class GoTestBuildType extends AbstractGoBuildType {
+		
+		public GoTestBuildType() {
+			super(BUILD_TYPE_BuildTests);
+		}
+		
+		@Override
+		protected String getBuildCommand() {
+			return "test -c";
+		}
+		
+		@Override
+		protected String getBinFilePath(GoPackageName goPackageName) throws CommonException {
+			return super.getBinFilePath(goPackageName) + ".test";
+		}
+		
+	}
+	
+	public static class GoRunTestsBuildType extends AbstractGoBuildType {
+		
+		public GoRunTestsBuildType() {
+			super(BUILD_TYPE_RunTests);
+		}
+		
+		@Override
+		protected String getBuildCommand() {
+			return "test";
+		}
+		
+		@Override
+		public Path getArtifactPath(BuildTarget buildTarget, IProject project) throws CommonException {
+			throw new CommonException("This configuration does not produce executable artifacts.");
+		}
+		
+	}
+	
+	public static class GoBuildTargetOperation extends CommonBuildTargetOperation {
+		
+		public GoBuildTargetOperation(BuildManager goBuildMgr, OperationInfo parentOpInfo, IProject project, 
+				Path buildToolPath, BuildTarget buildTarget, boolean fullBuild) {
+			super(goBuildMgr, parentOpInfo, project, buildToolPath, buildTarget, fullBuild);
+		}
 		
 		protected GoEnvironment goEnv;
 		protected Location sourceRootDir;
 		
-		public GoBuildTargetOperation(OperationInfo parentOpInfo, IProject project, Path buildToolPath,
-				BuildTarget buildTarget, boolean fullBuild) {
-			super(parentOpInfo, project, buildToolPath, buildTarget, fullBuild);
-		}
-		
 		@Override
 		public void execute(IProgressMonitor pm) throws CoreException, CommonException, OperationCancellation {
+			
+			ArrayList2<String> goBuildCmdLine = getToolCommandLine();
+			goBuildCmdLine.addElements(LaunchUtils.getParsedArguments(getEffectiveBuildOptions()));
 			
 			Location projectLocation = ResourceUtils.getProjectLocation(project);
 			
@@ -113,7 +217,7 @@ public class GoBuildManager extends BuildManager {
 				checkGoFilesInSourceRoot();
 			}
 			
-			ProcessBuilder pb = createBuildPB();
+			ProcessBuilder pb = goEnv.createProcessBuilder(goBuildCmdLine, sourceRootDir);
 			
 			ExternalProcessResult buildAllResult = runBuildTool(pm, pb);
 			doBuild_processBuildResult(buildAllResult);
@@ -126,16 +230,8 @@ public class GoBuildManager extends BuildManager {
 			
 			if(srcCheck.containsGoSources) {
 				LangCore.getToolManager().notifyMessage(StatusLevel.WARNING, "Go build: Warning!", 
-					MessageFormat.format(ERROR_SrcRootContainsGoFiles, sourceRootDir));
+					GoCoreMessages.ERROR_SrcRootContainsGoFiles(sourceRootDir));
 			}
-		}
-		
-		protected ProcessBuilder createBuildPB() throws CoreException, CommonException {
-			ArrayList2<String> goBuildCmdLine = getGoToolCommandLine();
-			goBuildCmdLine.addElements(
-				LaunchUtils.getParsedArguments(buildTarget.getBuildOptions()));
-			
-			return goEnv.createProcessBuilder(goBuildCmdLine, sourceRootDir);
 		}
 		
 		protected void doBuild_processBuildResult(ExternalProcessResult buildAllResult) 
@@ -161,7 +257,7 @@ public class GoBuildManager extends BuildManager {
 		}
 	}
 	
-	protected static ArrayList2<String> getGoToolCommandLine() throws CoreException {
+	protected static ArrayList2<String> getToolCommandLine() throws CoreException {
 		String compilerPath = GoEnvironmentPrefs.COMPILER_PATH.get();
 		
 		if(compilerPath.isEmpty()) {